docs(react-form): Update docs to reflect a change that was made pre-1.0.
Previously, the result of field.state.errors was an array of strings.
Now, it is an array of objects, from which you need to extract the
"message" values. This commit updates the TanStack Form docs for
React to reflect this. I am not sure of the correct syntax for Angular,
Solid, or Vue, but they will have the same issues.  Just search for
"errors.join" and this will locate the things that need to be changed.
Authoritative source for this change was a comment on my discussion post:
https://github.com/TanStack/form/discussions/1284#discussioncomment-12510456