feat(react-form): Add `withFieldGroup` (#1469)
* feat: add createFormGroup

* refactor: move createFormGroup to AppForm

While the previous separate implementation was compatible with AppForm, users
wouldn't have been able to use any field/form components
in the render itself. This commit allows them to do that,
at the expense of not being compatible with useForm.

* chore: add unit tests for createFormGroup types

* chore: add unit test for createFormGroup

* chore: export CreateFormGroupProps

* add DeepKeysOfType util type

This type will help with a lens API for forms
so that only names leading to the subset data
are allowed.

* feat: add initial FormLensApi draft

* chore: add FormLensApi tests batch

* fix(form-core): fix form.resetField() ignoring nested fields

* chore: complete form-core unit test for FormLensApi

* feat: add react adapter to form lens api

* fix: fix names for lens.Field and add test

* chore: export WithFormLensProps

* feat: add Subscribe and store to form lens

* feat: add mount method to FormLensApi

* fix: memoize innerProps to avoid focus loss in withFormLens

* refactor: use single useState instead of multiple useMemos

this is because useMemo is not intended for stable objects
see https://react.dev/reference/react/useMemo for more info

* feat: allow nesting withFormLenses

* remove createFormGroup for redundancy

* fix: widen typing of lens.Field/AppField to correct level

* docs: add withFormLens section

* fix: fix TName for lens component

* docs: fix typo in withFormLens

* feat: add lensErrors to FormLensApi store

* chore: adjust memo dependency in useFormLens

* chore: call userEvent.setup() in createFormHook tests

* refactor: move path concatenation to utils

* chore: move useLens to own file and rename

* feat: add FieldsMap and createFieldMap utils

* chore: migrate (most) lens references to field group

* chore: finalize migration from lens to field group

* ci: apply automated fixes and generate docs

* chore: remove accidental test file

* chore: add some unit tests for field mapping

* chore: add unit tests

* docs: update docs to use group

* docs: add caveat with field mapping

* docs: fix weird line break in alert text

* revert: remove FieldGroupApi#resetFieldMeta

the method appears to be a helper function for form.reset,
which is accessible from the field group API.
There does not seem to be a good reason to port this method
from FormApi.

* refactor: allow null or undefined for field group keys

users are able to do conditional rendering, but
TS generally doesn't pick up on it. Therefore,
while it is less type safe, it allows users to
use field groups in more locations than previously
possible.

* chore: add FieldGroupApi.test-d.ts

* docs(react-form): amend large form example with withFieldGroup

* chore(form-core): remove FieldGroupApi.reset

the reset is already accessible through group.form.reset.

* chore: fix broken unit test

---------

Co-authored-by: autofix-ci[bot] <114827586+autofix-ci[bot]@users.noreply.github.com>